Add stop_vms and lockfile options.
stop_vms: Stop the VM if is running and start it again after the backup. It may
be true or false and is false by default.

lockfile: Use a lock file to avoid two instances to run at the same time. It is
possible to set the file lock at the configuration file. If omitted the default
(/tmp/VmBackup.lock) is used.